iceberg
Module arrow

Conversion between Iceberg and Arrow schema

Structs

ArrowFileReader

ArrowFileReader is a wrapper around a FileRead that impls parquets AsyncFileReader.

ArrowReader

Reads data from Parquet files

ArrowReaderBuilder

Builder to create ArrowReader

Constants

DEFAULT_MAP_FIELD_NAME

When iceberg map type convert to Arrow map type, the default map field name is “key_value”.

UTC_TIME_ZONE

UTC time zone for Arrow timestamp type.

Traits

ArrowSchemaVisitor

A post order arrow schema visitor.

Functions

arrow_schema_to_schema

Convert Arrow schema to Iceberg schema.

arrow_struct_to_literal

Convert arrow struct array to iceberg struct value array. This function will assume the schema of arrow struct array is the same as iceberg struct type.

arrow_type_to_type

Convert Arrow type to iceberg type.

schema_to_arrow_schema

Convert iceberg schema to an arrow schema.

type_to_arrow_type

Convert iceberg type to an arrow type.