iceberg::arrow::delete_file_loader

Trait DeleteFileLoader

pub trait DeleteFileLoader {
    // Required method
    fn read_delete_file<'life0, 'life1, 'async_trait>(
        &'life0 self,
        task: &'life1 FileScanTaskDeleteFile,
        schema: SchemaRef,
    ) -> Pin<Box<dyn Future<Output = Result<ArrowRecordBatchStream>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Delete File Loader

Required Methods

fn read_delete_file<'life0, 'life1, 'async_trait>( &'life0 self, task: &'life1 FileScanTaskDeleteFile, schema: SchemaRef, ) -> Pin<Box<dyn Future<Output = Result<ArrowRecordBatchStream>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Read the delete file referred to in the task

Returns the contents of the delete file as a RecordBatch stream. Applies schema evolution.