iceberg::arrow

Function arrow_struct_to_literal

pub fn arrow_struct_to_literal(
    struct_array: &ArrayRef,
    ty: &StructType,
) -> Result<Vec<Option<Literal>>>

Convert arrow struct array to iceberg struct value array. This function will assume the schema of arrow struct array is the same as iceberg struct type.