iceberg::arrow

Trait StreamsInto

pub trait StreamsInto<R, S = ArrowRecordBatchStream> {
    // Required method
    fn stream(self, reader: R) -> Result<S>;
}

Trait indicating that the implementing type streams into a stream of type S using a reader of type R.

Required Methods

fn stream(self, reader: R) -> Result<S>

Stream from the reader and produce a stream of type S.