iceberg

Module io

File io implementation.

How to build FileIO

We provided a FileIOBuilder to build FileIO from scratch. For example:

use iceberg::Result;
use iceberg::io::FileIOBuilder;
use iceberg::io::storage::config::S3_REGION;

// Build a memory file io.
let file_io = FileIOBuilder::new("memory").build()?;
// Build an fs file io.
let file_io = FileIOBuilder::new("fs").build()?;
// Build an s3 file io.
let file_io = FileIOBuilder::new("s3")
    .with_prop(S3_REGION, "us-east-1")
    .build()?;

Or you can pass a path to ask FileIO to infer schema for you:

use iceberg::Result;
use iceberg::io::FileIO;
use iceberg::io::storage::config::S3_REGION;

// Build a memory file io.
let file_io = FileIO::from_path("memory:///")?.build()?;
// Build an fs file io.
let file_io = FileIO::from_path("fs:///tmp")?.build()?;
// Build an s3 file io.
let file_io = FileIO::from_path("s3://bucket/a")?
    .with_prop(S3_REGION, "us-east-1")
    .build()?;

How to use FileIO

Currently FileIO provides simple methods for file operations:

delete: Delete file.

exists: Check if file exists.

new_input: Create input file for reading.

new_output: Create output file for writing.

Re-exports

pub use storage::opendal::CustomAwsCredentialLoader;

pub use storage::opendal::OpenDalStorage;

pub use storage::opendal::OpenDalStorageFactory;

pub use storage::Storage;

pub use storage::StorageConfig;

pub use storage::StorageFactory;

Modules

storage

Storage interfaces for Iceberg.

Structs

Extensions

Container for storing type-safe extensions used to configure underlying FileIO behavior.

FileIO

FileIO implementation, used to manipulate files in underlying storage.

FileIOBuilder

Builder for FileIO.

FileMetadata

The struct the represents the metadata of a file.

InputFile

Input file is used for reading from files.

MetadataLocation

Newtype wrapper for the metadata location string, used as an extension so that RefreshableOpenDalStorage can pass it to load_credentials.

OutputFile

Output file is used for writing to files..

StorageCredential

Storage credentials for accessing cloud storage.

Traits

FileRead

Trait for reading file.

FileWrite

Trait for writing file.

StorageCredentialsLoader

Trait for loading storage credentials dynamically.