iceberg::spec

Struct TableMetadata

pub struct TableMetadata { /* private fields */ }

Fields for the version 2 of the table metadata.

We assume that this data structure is always valid, so we will panic when invalid error happens. We check the validity of this data structure when constructing.

Implementations

impl TableMetadata

pub fn into_builder( self, current_file_location: Option<String>, ) -> TableMetadataBuilder

Convert this Table Metadata into a builder for modification.

current_file_location is the location where the current version of the metadata file is stored. This is used to update the metadata log. If current_file_location is None, the metadata log will not be updated. This should only be used to stage-create tables.

pub fn format_version(&self) -> FormatVersion

Returns format version of this metadata.

pub fn uuid(&self) -> Uuid

Returns uuid of current table.

pub fn location(&self) -> &str

Returns table location.

pub fn last_sequence_number(&self) -> i64

Returns last sequence number.

pub fn next_sequence_number(&self) -> i64

Returns the next sequence number for the table.

For format version 1, it always returns the initial sequence number. For other versions, it returns the last sequence number incremented by 1.

pub fn last_column_id(&self) -> i32

Returns the last column id.

pub fn last_partition_id(&self) -> i32

Returns the last partition_id

pub fn last_updated_timestamp(&self) -> Result<DateTime<Utc>>

Returns last updated time.

pub fn last_updated_ms(&self) -> i64

Returns last updated time in milliseconds.

pub fn schemas_iter(&self) -> impl ExactSizeIterator<Item = &SchemaRef>

Returns schemas

pub fn schema_by_id(&self, schema_id: SchemaId) -> Option<&SchemaRef>

Lookup schema by id.

pub fn current_schema(&self) -> &SchemaRef

Get current schema

pub fn current_schema_id(&self) -> SchemaId

Get the id of the current schema

pub fn partition_specs_iter( &self, ) -> impl ExactSizeIterator<Item = &PartitionSpecRef>

Returns all partition specs.

pub fn partition_spec_by_id(&self, spec_id: i32) -> Option<&PartitionSpecRef>

Lookup partition spec by id.

pub fn default_partition_spec(&self) -> &PartitionSpecRef

Get default partition spec

pub fn default_partition_type(&self) -> &StructType

Return the partition type of the default partition spec.

pub fn default_partition_spec_id(&self) -> i32

Returns spec id of the “current” partition spec.

pub fn snapshots(&self) -> impl ExactSizeIterator<Item = &SnapshotRef>

Returns all snapshots

pub fn snapshot_by_id(&self, snapshot_id: i64) -> Option<&SnapshotRef>

Lookup snapshot by id.

pub fn history(&self) -> &[SnapshotLog]

Returns snapshot history.

pub fn metadata_log(&self) -> &[MetadataLog]

Returns the metadata log.

pub fn current_snapshot(&self) -> Option<&SnapshotRef>

Get current snapshot

pub fn current_snapshot_id(&self) -> Option<i64>

Get the current snapshot id

pub fn snapshot_for_ref(&self, ref_name: &str) -> Option<&SnapshotRef>

Get the snapshot for a reference Returns an option if the ref_name is not found

pub fn sort_orders_iter(&self) -> impl ExactSizeIterator<Item = &SortOrderRef>

Return all sort orders.

pub fn sort_order_by_id(&self, sort_order_id: i64) -> Option<&SortOrderRef>

Lookup sort order by id.

pub fn default_sort_order(&self) -> &SortOrderRef

Returns default sort order id.

pub fn default_sort_order_id(&self) -> i64

Returns default sort order id.

pub fn properties(&self) -> &HashMap<String, String>

Returns properties of table.

pub fn statistics_iter(&self) -> impl ExactSizeIterator<Item = &StatisticsFile>

Return location of statistics files.

pub fn partition_statistics_iter( &self, ) -> impl ExactSizeIterator<Item = &PartitionStatisticsFile>

Return location of partition statistics files.

pub fn statistics_for_snapshot( &self, snapshot_id: i64, ) -> Option<&StatisticsFile>

Get a statistics file for a snapshot id.

pub fn partition_statistics_for_snapshot( &self, snapshot_id: i64, ) -> Option<&PartitionStatisticsFile>

Get a partition statistics file for a snapshot id.

pub fn encryption_keys_iter( &self, ) -> impl ExactSizeIterator<Item = (&String, &String)>

Iterate over all encryption keys

pub fn encryption_key(&self, key_id: &str) -> Option<&String>

Get the encryption key for a given key id

pub async fn read_from( file_io: &FileIO, metadata_location: impl AsRef<str>, ) -> Result<TableMetadata>

Read table metadata from the given location.

pub async fn write_to( &self, file_io: &FileIO, metadata_location: impl AsRef<str>, ) -> Result<()>

Write table metadata to the given location.