iceberg::spec

Struct ViewMetadataBuilder

pub struct ViewMetadataBuilder { /* private fields */ }

Manipulating view metadata.

For this builder the order of called functions matters. All operations applied to the ViewMetadata are tracked in changes as a chronologically ordered vec of ViewUpdate. If an operation does not lead to a change of the ViewMetadata, the corresponding update is omitted from changes.

Implementations

impl ViewMetadataBuilder

pub fn new( location: String, schema: Schema, view_version: ViewVersion, format_version: ViewFormatVersion, properties: HashMap<String, String>, ) -> Result<Self>

Creates a new view metadata builder.

pub fn new_from_metadata(previous: ViewMetadata) -> Self

Creates a new view metadata builder from the given metadata to modify it.

pub fn from_view_creation(view_creation: ViewCreation) -> Result<Self>

Creates a new view metadata builder from the given view creation.

pub fn upgrade_format_version( self, format_version: ViewFormatVersion, ) -> Result<Self>

Upgrade FormatVersion. Downgrades are not allowed.

Errors

Cannot downgrade to older format versions.

pub fn set_location(self, location: String) -> Self

Set the location of the view, stripping any trailing slashes.

pub fn set_current_version_id(self, version_id: i32) -> Result<Self>

Set an existing view version as the current version.

Errors

The specified version_id does not exist.

The specified version_id is -1 but no version has been added.

pub fn set_current_version( self, view_version: ViewVersion, schema: Schema, ) -> Result<Self>

Add a new view version and set it as current.

pub fn add_version(self, view_version: ViewVersion) -> Result<Self>

Add a new version to the view.

Errors

The schema ID of the version is set to -1, but no schema has been added.

The schema ID of the specified version is unknown.

Multiple queries for the same dialect are added.

pub fn add_schema(self, schema: Schema) -> Self

Add a new schema to the view.

pub fn set_properties(self, updates: HashMap<String, String>) -> Result<Self>

Update properties of the view.

pub fn remove_properties(self, removals: &[String]) -> Self

Remove properties from the view

pub fn assign_uuid(self, uuid: Uuid) -> Self

Assign a new UUID to the view.

pub fn build(self) -> Result<ViewMetadataBuildResult>

Build the ViewMetadata from the changes.