iceberg::transaction

Trait ApplyTransactionAction

pub trait ApplyTransactionAction {
    // Required method
    fn apply(self, tx: Transaction) -> Result<Transaction>;
}

A helper trait for applying a TransactionAction to a Transaction.

This is implemented for all TransactionAction types to allow easy chaining of actions into a transaction context.

Required Methods

fn apply(self, tx: Transaction) -> Result<Transaction>

Adds this action to the given transaction.

Arguments

tx - The transaction to apply the action to.

Returns

The modified transaction containing this action, or an error if the operation fails.