Module iceberg::writer

Iceberg writer module.

This module contains the generic writer trait and specific writer implementation. We categorize the writer into two types:

FileWriter: writer for physical file format (Such as parquet, orc).

IcebergWriter: writer for logical format provided by iceberg table (Such as data file, equality delete file, position delete file) or other function (Such as partition writer, delta writer).

The IcebergWriter will use FileWriter to write underly physical file.

We hope the writer interface can be extensible and flexible. Each writer can be create config independently and combined together to build a writer which have complex write logic. E.g. combine FanoutPartitionWriter, DataFileWriter, ParquetWriter to get a writer can split the data automatelly according to partition and write down as parquet physical format.

For this purpose, there are four trait corresponding to these writer:

IcebergWriterBuilder

IcebergWriter

FileWriterBuilder

FileWriter

User can create specific writer builder, combine them and build the writer finally. Also user can custom own writer and implement writer trait for them so that the custom writer can integrate with existing writer. (See following example)

Simple example for the data file writer used parquet physical format:

use std::sync::Arc;

use arrow_array::{ArrayRef, BooleanArray, Int32Array, RecordBatch, StringArray};
use async_trait::async_trait;
use iceberg::io::{FileIO, FileIOBuilder};
use iceberg::spec::DataFile;
use iceberg::transaction::Transaction;
use iceberg::writer::base_writer::data_file_writer::DataFileWriterBuilder;
use iceberg::writer::file_writer::location_generator::{
    DefaultFileNameGenerator, DefaultLocationGenerator,
};
use iceberg::writer::file_writer::ParquetWriterBuilder;
use iceberg::writer::{IcebergWriter, IcebergWriterBuilder};
use iceberg::{Catalog, Result, TableIdent};
use iceberg_catalog_memory::MemoryCatalog;
use parquet::file::properties::WriterProperties;
#[tokio::main]
async fn main() -> Result<()> {
    // Build your file IO.
    let file_io = FileIOBuilder::new("memory").build()?;
    // Connect to a catalog.
    let catalog = MemoryCatalog::new(file_io, None);
    // Load table from catalog.
    let table = catalog
        .load_table(&TableIdent::from_strs(["hello", "world"])?)
        .await?;
    let location_generator = DefaultLocationGenerator::new(table.metadata().clone()).unwrap();
    let file_name_generator = DefaultFileNameGenerator::new(
        "test".to_string(),
        None,
        iceberg::spec::DataFileFormat::Parquet,
    );

    // Create a parquet file writer builder. The parameter can get from table.
    let parquet_writer_builder = ParquetWriterBuilder::new(
        WriterProperties::default(),
        table.metadata().current_schema().clone(),
        table.file_io().clone(),
        location_generator.clone(),
        file_name_generator.clone(),
    );
    // Create a data file writer using parquet file writer builder.
    let data_file_writer_builder = DataFileWriterBuilder::new(parquet_writer_builder, None, 0);
    // Build the data file writer
    let mut data_file_writer = data_file_writer_builder.build().await.unwrap();

    // Write the data using data_file_writer...

    // Close the write and it will return data files back
    let data_files = data_file_writer.close().await.unwrap();

    Ok(())
}

Custom writer to record latency

use std::time::Instant;

use arrow_array::RecordBatch;
use iceberg::io::FileIOBuilder;
use iceberg::spec::DataFile;
use iceberg::writer::base_writer::data_file_writer::DataFileWriterBuilder;
use iceberg::writer::file_writer::location_generator::{
    DefaultFileNameGenerator, DefaultLocationGenerator,
};
use iceberg::writer::file_writer::ParquetWriterBuilder;
use iceberg::writer::{IcebergWriter, IcebergWriterBuilder};
use iceberg::{Catalog, Result, TableIdent};
use iceberg_catalog_memory::MemoryCatalog;
use parquet::file::properties::WriterProperties;

#[derive(Clone)]
struct LatencyRecordWriterBuilder<B> {
    inner_writer_builder: B,
}

impl<B: IcebergWriterBuilder> LatencyRecordWriterBuilder<B> {
    pub fn new(inner_writer_builder: B) -> Self {
        Self {
            inner_writer_builder,
        }
    }
}

#[async_trait::async_trait]
impl<B: IcebergWriterBuilder> IcebergWriterBuilder for LatencyRecordWriterBuilder<B> {
    type R = LatencyRecordWriter<B::R>;

    async fn build(self) -> Result<Self::R> {
        Ok(LatencyRecordWriter {
            inner_writer: self.inner_writer_builder.build().await?,
        })
    }
}
struct LatencyRecordWriter<W> {
    inner_writer: W,
}

#[async_trait::async_trait]
impl<W: IcebergWriter> IcebergWriter for LatencyRecordWriter<W> {
    async fn write(&mut self, input: RecordBatch) -> Result<()> {
        let start = Instant::now();
        self.inner_writer.write(input).await?;
        let _latency = start.elapsed();
        // record latency...
        Ok(())
    }

    async fn close(&mut self) -> Result<Vec<DataFile>> {
        let start = Instant::now();
        let res = self.inner_writer.close().await?;
        let _latency = start.elapsed();
        // record latency...
        Ok(res)
    }
}

#[tokio::main]
async fn main() -> Result<()> {
    // Build your file IO.
    let file_io = FileIOBuilder::new("memory").build()?;
    // Connect to a catalog.
    let catalog = MemoryCatalog::new(file_io, None);
    // Load table from catalog.
    let table = catalog
        .load_table(&TableIdent::from_strs(["hello", "world"])?)
        .await?;
    let location_generator = DefaultLocationGenerator::new(table.metadata().clone()).unwrap();
    let file_name_generator = DefaultFileNameGenerator::new(
        "test".to_string(),
        None,
        iceberg::spec::DataFileFormat::Parquet,
    );

    // Create a parquet file writer builder. The parameter can get from table.
    let parquet_writer_builder = ParquetWriterBuilder::new(
        WriterProperties::default(),
        table.metadata().current_schema().clone(),
        table.file_io().clone(),
        location_generator.clone(),
        file_name_generator.clone(),
    );
    // Create a data file writer builder using parquet file writer builder.
    let data_file_writer_builder = DataFileWriterBuilder::new(parquet_writer_builder, None, 0);
    // Create latency record writer using data file writer builder.
    let latency_record_builder = LatencyRecordWriterBuilder::new(data_file_writer_builder);
    // Build the final writer
    let mut latency_record_data_file_writer = latency_record_builder.build().await.unwrap();

    Ok(())
}

Modules

base_writer

Base writer module contains the basic writer provide by iceberg: DataFileWriter, PositionDeleteFileWriter, EqualityDeleteFileWriter.

file_writer

This module contains the writer for data file format supported by iceberg: parquet, orc.

Traits

CurrentFileStatus

The current file status of iceberg writer. It implement for the writer which write a single file.

IcebergWriter

The iceberg writer used to write data to iceberg table.

IcebergWriterBuilder

The builder for iceberg writer.