iceberg::writer

Module file_writer

This module contains the writer for data file format supported by iceberg: parquet, orc.

Modules

location_generator

This module contains the location generator and file name generator for generating path of data file.

rolling_writer

Module providing writers that can automatically roll over to new files based on size thresholds.

Structs

ParquetWriter

`ParquetWriter`` is used to write arrow data into parquet file on storage.

ParquetWriterBuilder

ParquetWriterBuilder is used to builder a ParquetWriter

Traits

FileWriter

File writer focus on writing record batch to different physical file format.(Such as parquet. orc)

FileWriterBuilder

File writer builder trait.