iceberg::spec

Struct TableProperties

pub struct TableProperties {
    pub commit_num_retries: usize,
    pub commit_min_retry_wait_ms: u64,
    pub commit_max_retry_wait_ms: u64,
    pub commit_total_retry_timeout_ms: u64,
    pub write_format_default: String,
    pub write_target_file_size_bytes: usize,
}

TableProperties that contains the properties of a table.

Fields
commit_num_retries: usize

The number of times to retry a commit.

commit_min_retry_wait_ms: u64

The minimum wait time between retries.

commit_max_retry_wait_ms: u64

The maximum wait time between retries.

commit_total_retry_timeout_ms: u64

The total timeout for commit retries.

write_format_default: String

The default format for files.

write_target_file_size_bytes: usize

The target file size for files.

Implementations

impl TableProperties

pub const PROPERTY_FORMAT_VERSION: &str = "format-version"

Reserved table property for table format version.

Iceberg will default a new table’s format version to the latest stable and recommended version. This reserved property keyword allows users to override the Iceberg format version of the table metadata.

If this table property exists when creating a table, the table will use the specified format version. If a table updates this property, it will try to upgrade to the specified format version.

pub const PROPERTY_UUID: &str = "uuid"

Reserved table property for table UUID.

pub const PROPERTY_SNAPSHOT_COUNT: &str = "snapshot-count"

Reserved table property for the total number of snapshots.

pub const PROPERTY_CURRENT_SNAPSHOT_SUMMARY: &str = "current-snapshot-summary"

Reserved table property for current snapshot summary.

pub const PROPERTY_CURRENT_SNAPSHOT_ID: &str = "current-snapshot-id"

Reserved table property for current snapshot id.

pub const PROPERTY_CURRENT_SNAPSHOT_TIMESTAMP: &str = "current-snapshot-timestamp-ms"

Reserved table property for current snapshot timestamp.

pub const PROPERTY_CURRENT_SCHEMA: &str = "current-schema"

Reserved table property for the JSON representation of current schema.

pub const PROPERTY_DEFAULT_PARTITION_SPEC: &str = "default-partition-spec"

Reserved table property for the JSON representation of current(default) partition spec.

pub const PROPERTY_DEFAULT_SORT_ORDER: &str = "default-sort-order"

Reserved table property for the JSON representation of current(default) sort order.

pub const PROPERTY_METADATA_PREVIOUS_VERSIONS_MAX: &str = "write.metadata.previous-versions-max"

Property key for max number of previous versions to keep.

pub const PROPERTY_METADATA_PREVIOUS_VERSIONS_MAX_DEFAULT: usize = 100usize

Default value for max number of previous versions to keep.

pub const PROPERTY_WRITE_PARTITION_SUMMARY_LIMIT: &str = "write.summary.partition-limit"

Property key for max number of partitions to keep summary stats for.

pub const PROPERTY_WRITE_PARTITION_SUMMARY_LIMIT_DEFAULT: u64 = 0u64

Default value for the max number of partitions to keep summary stats for.

pub const RESERVED_PROPERTIES: [&str; 9]

Reserved Iceberg table properties list.

Reserved table properties are only used to control behaviors when creating or updating a table. The value of these properties are not persisted as a part of the table metadata.

pub const PROPERTY_COMMIT_NUM_RETRIES: &str = "commit.retry.num-retries"

Property key for number of commit retries.

pub const PROPERTY_COMMIT_NUM_RETRIES_DEFAULT: usize = 4usize

Default value for number of commit retries.

pub const PROPERTY_COMMIT_MIN_RETRY_WAIT_MS: &str = "commit.retry.min-wait-ms"

Property key for minimum wait time (ms) between retries.

pub const PROPERTY_COMMIT_MIN_RETRY_WAIT_MS_DEFAULT: u64 = 100u64

Default value for minimum wait time (ms) between retries.

pub const PROPERTY_COMMIT_MAX_RETRY_WAIT_MS: &str = "commit.retry.max-wait-ms"

Property key for maximum wait time (ms) between retries.

pub const PROPERTY_COMMIT_MAX_RETRY_WAIT_MS_DEFAULT: u64 = 60_000u64

Default value for maximum wait time (ms) between retries.

pub const PROPERTY_COMMIT_TOTAL_RETRY_TIME_MS: &str = "commit.retry.total-timeout-ms"

Property key for total maximum retry time (ms).

pub const PROPERTY_COMMIT_TOTAL_RETRY_TIME_MS_DEFAULT: u64 = 1_800_000u64

Default value for total maximum retry time (ms).

pub const PROPERTY_DEFAULT_FILE_FORMAT: &str = "write.format.default"

Default file format for data files

pub const PROPERTY_DELETE_DEFAULT_FILE_FORMAT: &str = "write.delete.format.default"

Default file format for delete files

pub const PROPERTY_DEFAULT_FILE_FORMAT_DEFAULT: &str = "parquet"

Default value for data file format

pub const PROPERTY_WRITE_TARGET_FILE_SIZE_BYTES: &str = "write.target-file-size-bytes"

Target file size for newly written files.

pub const PROPERTY_WRITE_TARGET_FILE_SIZE_BYTES_DEFAULT: usize = 536_870_912usize

Default target file size