Module iceberg::writer::file_writer::location_generator

This module contains the location generator and file name generator for generating path of data file.

Structs

DefaultFileNameGenerator

DefaultFileNameGenerator used to generate file name for data file. The file name can be passed to LocationGenerator to generate the location of the file. The file name format is “{prefix}-{file_count}[-{suffix}].{file_format}”.

DefaultLocationGenerator

DefaultLocationGenerator used to generate the data dir location of data file. The location is generated based on the table location and the data location in table properties.

Traits

FileNameGenerator

FileNameGeneratorTrait used to generate file name for data file. The file name can be passed to LocationGenerator to generate the location of the file.

LocationGenerator

LocationGenerator used to generate the location of data file.